requirements.txt
requests>=2.25.1
paramiko>=2.7.0
pyyaml>=5.4.1
